DDP-5667: Move activityAnswer.component.ts to a separate directory
