+++ ddp-workspace/projects/ddp-sdk/src/lib/components/activityForm/answers/activity-answer/activityAnswer.component.ts
@@ -1,9 +1,9 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { AbstractActivityQuestionBlock } from '../../../models/activity/abstractActivityQuestionBlock';
-import { AnswerValue } from '../../../models/activity/answerValue';
-import { QuestionType } from '../../../models/activity/questionType';
-import { BlockType } from '../../../models/activity/blockType';
-import { BlockVisibility } from '../../../models/activity/blockVisibility';
+import { AbstractActivityQuestionBlock } from '../../../../models/activity/abstractActivityQuestionBlock';
+import { AnswerValue } from '../../../../models/activity/answerValue';
+import { QuestionType } from '../../../../models/activity/questionType';
+import { BlockType } from '../../../../models/activity/blockType';
+import { BlockVisibility } from '../../../../models/activity/blockVisibility';
 
 @Component({
     selector: 'ddp-activity-answer',
