+++ ddp-workspace/projects/ddp-sdk/src/public-api.ts
@@ -128,7 +128,7 @@ export * from './lib/components/activityForm/activity-blocks/activityBlock/activ
 export * from './lib/components/activityForm/answers/activityAgreementAnswer.component';
 export * from './lib/components/activityForm/answers/activityCompositeAnswer.component';
 export * from './lib/components/activityForm/answers/activityDateAnswer.component';
-export * from './lib/components/activityForm/answers/activityAnswer.component';
+export * from './lib/components/activityForm/answers/activity-answer/activityAnswer.component';
 export * from './lib/components/activityForm/answers/activityNumericAnswer.component';
 export * from './lib/components/activityForm/answers/activityEmailInput.component';
 export * from './lib/components/activityForm/answers/activityTextAnswer.component';
